+++ frontend/app/components/videoComponents/ControllerBar/styles.ts
