+++ frontend/app/components/videoComponents/VideoNavigator/style.ts
