+++ frontend/app/components/videoComponents/VideoPlayer/styles.ts
