+++ frontend/app/layout.tsx
@@ -17,8 +17,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "BrainrotTube",
+  description: "BrainrotTube is a video streaming platform for brainrot content.",
 };
 
 export default function RootLayout({
