+++ frontend/app/page.tsx
@@ -1,4 +1,4 @@
-import VideoContent from './components/videoContent/VideoContent';
+import VideoContent from './components/videoComponents/VideoContent';
 import React from 'react';
 
 export default function Home() {
